Adding feed forward network to blocks
